Move numeric range parsing logic
* Add NumericRangeSpec::fromString by moving parsing from QueryTermSimple
* use the new fromString in QueryTermSimple constructor
