+++ searchlib/src/vespa/searchlib/query/CMakeLists.txt
@@ -2,6 +2,7 @@
 vespa_add_library(searchlib_query OBJECT
     SOURCES
     from_proto.cpp
+    numeric_range_spec.cpp
     query_normalization.cpp
     query_stack_iterator.cpp
     query_term_decoder.cpp
