+++ searchlib/src/vespa/searchlib/query/query_term_simple.cpp
@@ -2,12 +2,10 @@
 
 #include "query_term_simple.h"
 #include "base.h"
-#include <vespa/vespalib/locale/c.h>
 #include <vespa/vespalib/objects/visit.h>
 #include <vespa/vespalib/util/classname.h>
 #include <cmath>
 #include <limits>
-#include <charconv>
 
 using search::NumericRangeSpec;
 
@@ -25,171 +23,23 @@ constexpr bool isRepresentableByInt64(double d) noexcept {
             (d < double(std::numeric_limits<int64_t>::max()));
 }
 
-bool isFullRange(std::string_view s) noexcept {
-    const size_t sz(s.size());
-    return (sz >= 3u) &&
-            (s[0] == '<' || s[0] == '[') &&
-            (s[sz-1] == '>' || s[sz-1] == ']');
-}
-
-bool isPartialRange(std::string_view s) noexcept {
-    return (s.size() > 1) &&
-            ((s[0] == '<') || (s[0] == '>'));
-}
-
 constexpr int64_t NEG_MIN_I64 = std::numeric_limits<int64_t>::min();
 constexpr int64_t POS_MAX_I64 = std::numeric_limits<int64_t>::max();
 
-bool parseNumberAsInt64(const char *startp, const char *endptr, int64_t &t2) {
-    if (*startp == '+')
-        ++startp;
-    auto result = std::from_chars(startp, endptr, t2);
-    if (result.ec == std::errc{} && result.ptr == endptr) {
+bool isRange(std::string_view s) noexcept {
+    if (s.size() < 2) return false;
+    // Check for partial range: <value or >value
+    if ((s[0] == '<') || (s[0] == '>')) {
+        return true;
+    }
+    // Check for full range: [value;value] or <value;value>
+    const size_t sz(s.size());
+    if ((sz >= 3u) &&
+        (s[0] == '<' || s[0] == '[') &&
+        (s[sz-1] == '>' || s[sz-1] == ']')) {
         return true;
     }
     return false;
-}
-
-bool parseNumberAsDouble(const char *startp, const char *endptr, double &target) {
-    if (startp == endptr) return false;
-    char *parsed_to;
-    double dv = vespalib::locale::c::strtod(startp, &parsed_to);
-    if (parsed_to != endptr) [[unlikely]] {
-        return false;
-    }
-    if (std::isnan(dv)) [[unlikely]] {
-        return false;
-    }
-    target = dv;
-    return true;
-}
-
-struct TwiceParser {
-    bool invalid = true;
-    bool valid_i = false;
-    double d = 0.0;
-    int64_t i = 0;
-    TwiceParser(std::string_view view) {
-        const char *s = view.data();
-        const char *e = view.data() + view.size();
-        if (parseNumberAsDouble(s, e, d)) {
-            invalid = false;
-            valid_i = parseNumberAsInt64(s, e, i);
-        }
-    }
-};
-
-std::unique_ptr<NumericRangeSpec> parsePartialRange(const std::string_view term) {
-    auto result = std::make_unique<NumericRangeSpec>();
-    TwiceParser parsed(term.substr(1));
-    if (parsed.invalid) {
-        return {};
-    }
-    if (term[0] == '<') {
-        result->upper_inclusive = false;
-        result->fp_upper_limit = parsed.d;
-        result->int64_upper_limit = parsed.i;
-        result->valid_integers = parsed.valid_i;
-        result->valid = true;
-        return result;
-    }
-    if (term[0] == '>') {
-        result->lower_inclusive = false;
-        result->fp_lower_limit = parsed.d;
-        result->int64_lower_limit = parsed.i;
-        result->valid_integers = parsed.valid_i;
-        result->valid = true;
-        return result;
-    }
-    return {};
-}
-
-std::unique_ptr<NumericRangeSpec> parseNoRange(const std::string &term) {
-    TwiceParser parsed(term);
-    if (parsed.invalid) return {};
-    auto result = std::make_unique<NumericRangeSpec>();
-    result->lower_inclusive = true;
-    result->upper_inclusive = true;
-    result->fp_lower_limit = parsed.d;
-    result->fp_upper_limit = parsed.d;
-    result->valid_integers = parsed.valid_i;
-    result->int64_lower_limit = parsed.i;
-    result->int64_upper_limit = parsed.i;
-    result->valid = true;
-    return result;
-}
-
-std::unique_ptr<NumericRangeSpec> parseFullRange(const std::string &_term) {
-    auto result = std::make_unique<NumericRangeSpec>();
-    std::string_view rest(_term.c_str() + 1, _term.size() - 2);
-    std::string_view parts[9];
-    size_t numParts(0);
-    while (! rest.empty() && ((numParts + 1) < NELEMS(parts))) {
-        size_t pos(rest.find(';'));
-        if (pos != std::string::npos) {
-            parts[numParts++] = rest.substr(0, pos);
-            rest = rest.substr(pos + 1);
-            if (rest.empty()) {
-                parts[numParts++] = rest;
-            }
-        } else {
-            parts[numParts++] = rest;
-            rest = std::string_view();
-        }
-    }
-    if (numParts < 2) return {};
-    if (NELEMS(parts) <= numParts) return {};
-
-    result->lower_inclusive = (_term[0] == '[');
-    result->upper_inclusive = (_term[_term.size() - 1] == ']');
-
-    bool valid_i = true;
-    if (parts[0].empty()) {
-        // note empty -> no limit
-        result->lower_inclusive = true; // <;3] is same as [;3]
-    } else {
-        TwiceParser parsed(parts[0]);
-        if (parsed.invalid) return {};
-        valid_i = parsed.valid_i;
-        result->fp_lower_limit = parsed.d;
-        result->int64_lower_limit = parsed.i;
-    }
-    if (parts[1].empty()) {
-        // note empty -> no limit
-        result->upper_inclusive = true; // [3;> is same as [3;]
-    } else {
-        TwiceParser parsed(parts[1]);
-        if (parsed.invalid) return {};
-        valid_i = valid_i && parsed.valid_i;
-        result->fp_upper_limit = parsed.d;
-        result->int64_upper_limit = parsed.i;
-    }
-    result->valid_integers = valid_i;
-    if (numParts > 2) {
-        result->rangeLimit = static_cast<int32_t>(strtol(parts[2].data(), nullptr, 0));
-        if (numParts > 3) {
-            if (numParts < 5) {
-                return {};
-            }
-            result->diversityAttribute = parts[3];
-            result->maxPerGroup = strtoul(parts[4].data(), nullptr, 0);
-            if ((result->maxPerGroup > 0) && (numParts > 5)) {
-                char *err = nullptr;
-                size_t cutoffGroups = strtoul(parts[5].data(), &err, 0);
-                if ((err == nullptr) || (size_t(err - parts[5].data()) == parts[5].size())) {
-                    result->diversityCutoffGroups = cutoffGroups;
-                }
-                if (numParts > 6) {
-                    result->diversityCutoffStrict = (parts[6] == "strict");
-                    if (numParts > 7) {
-                        return {};
-                    }
-                }
-            }
-        }
-    }
-    result->valid = true;
-    return result;
 }
 
 } // namespace
@@ -382,14 +232,9 @@ QueryTermSimple::QueryTermSimple(const string & term_, Type type)
     _fuzzy_max_edit_distance(2),
     _fuzzy_prefix_lock_length(0)
 {
-    if (isFullRange(_term)) {
-        _numeric_range = parseFullRange(_term);
+    _numeric_range = NumericRangeSpec::fromString(_term);
+    if (isRange(_term)) {
         _valid = bool(_numeric_range);
-    } else if (isPartialRange(_term)) {
-        _numeric_range = parsePartialRange(_term);
-        _valid = bool(_numeric_range);
-    } else {
-        _numeric_range = parseNoRange(_term);
     }
 }
